Extract command execution builtin
