+++ src/builtins/index.ts
@@ -1,1 +1,2 @@
+export * from './command-execution.builtin';
 export * from './standard-setup';
